chore: Move project name domain to project sidebar
