+++ web_ui/src/pages/project-details/components/project-sidebar/project-name-domain/project-name-domain.module.scss
